[CG Charter] test suite paragraph was for WGs
@ianbjacobs @wseltzer
The test suite section for the CG Charter template points to a page that is only about WGs.  It includes the 2 ways to include test suites in a WG (either as part of the spec or not) and also directly refers to the W3C patent policy.  It's not at all relevant to CGs. 

It would be better just to point to the GitHub LICENSE.md file if in GitHub.

 I followed links in that material until I found the W3C license used and proposed putting that in the LICENSE.md file for the WebVR CG.   That's a separate change I'll suggest in the W3C LICENSE.md file for CGs.